feat(rhino): bim elements commands
