+++ ConnectorRhino/ConnectorRhino/BIMElements/BIMElementFilter.cs
