+++ ConnectorRhino/ConnectorRhino/ConnectorRhino.csproj
@@ -48,8 +48,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConnectorRhinoCommand.cs" />
-    <Compile Include="SchemaBuilder\SchemaBuilderCommand.cs" />
-    <Compile Include="SchemaBuilder\SchemaObjectFilter.cs" />
+    <Compile Include="BIMElements\BIMElementCommands.cs" />
+    <Compile Include="BIMElements\BIMElementFilter.cs" />
     <Compile Include="UI\ConnectorBindingsRhino.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils.cs" />
